Move Storable definition and derive to ink_storage_traits
