--- crates/primitives/derive/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "ink_primitives_derive"
-version = "4.0.0-alpha.1"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-
-license = "Apache-2.0"
-readme = "../README.md"
-repository = "https://github.com/paritytech/ink"
-documentation = "https://docs.rs/ink_primitives_derive"
-homepage = "https://www.parity.io/"
-description = "[ink!] Derive macros for common ink_primitives defined traits."
-keywords = ["wasm", "parity", "webassembly", "blockchain", "edsl"]
-categories = ["no-std", "embedded"]
-include = ["Cargo.toml", "src/**/*.rs", "README.md", "LICENSE"]
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1"
-syn = { version = "1", features = ["full"] }
-proc-macro2 = "1"
-synstructure = "0.12.4"
-
-[dev-dependencies]
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive", "full"] }
-ink_env = { version = "4.0.0-alpha.1", path = "../../env" }
-ink_primitives = { version = "4.0.0-alpha.1", path = ".." }
-ink_metadata = { version = "4.0.0-alpha.1", path = "../../metadata" }
-ink_prelude = { version = "4.0.0-alpha.1", path = "../../prelude/" }
